requirements.txt
anyio
appnope
argon2-cffi
argon2-cffi-bindings
asttokens
async-lru
attrs
Babel
backcall
beautifulsoup4
bidict==0.22.1
bleach
Bottleneck
Brotli
certifi
cffi
charset-normalizer
click
comm
cryptography
debugpy
decorator
defusedxml
distro
exceptiongroup
executing
fastjsonschema
Flask
Flask-SocketIO==5.3.6
greenlet
h11==0.14.0
idna
importlib-metadata
ipykernel
ipython
itsdangerous
jedi
Jinja2
JPype1==1.4.1
json5
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab_server
MarkupSafe
matplotlib-inline
mistune
mkl-fft
mkl-random
mkl-service==2.4.0
mysqlclient
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numexpr
numpy
overrides
packaging
pandas
pandocfilters
parso
pexpect
pickleshare
platformdirs
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
Pygments
PyMySQL
pyOpenSSL
PyPDF2==2.10.5
PySocks
python-dateutil
python-engineio==4.8.0
python-json-logger
python-socketio==5.10.0
pytz
PyYAML
pyzmq
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rpds-py
Send2Trash
simple-websocket==1.0.0
six
sniffio
soupsieve
SQLAlchemy
stack-data
tabula-py==2.9.0
terminado
tinycss2
tomli
tornado
tqdm==4.66.1
traitlets
typing_extensions
tzdata
urllib3
wcwidth
webencodings==0.5.1
websocket-client
Werkzeug
wsproto==1.2.0
zipp
